backend/requirements.txt
# =====================================================
# CORE DEPENDENCIES
# =====================================================
Flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
Werkzeug>=2.2.0
itsdangerous>=2.1.0

# =====================================================
# DATABASE
# =====================================================
SQLAlchemy>=2.0.0
pymysql>=1.1.0
mysql-connector-python>=8.0.33

# =====================================================
# PRODUCTION WSGI SERVER
# =====================================================
gunicorn>=20.1.0

# =====================================================
# ENVIRONMENT & CONFIG
# =====================================================
python-dotenv>=0.19.0

# =====================================================
# SECURITY & ENCRYPTION
# =====================================================
cryptography>=40.0.0
Werkzeug>=2.2.0

# =====================================================
# VALIDATION & API
# =====================================================
pydantic>=1.9.0
email-validator>=1.2.0

# =====================================================
# UTILITIES
# =====================================================
flask-limiter>=3.3.0
